Create Program10
This program will take input from the user, in the form of mm/dd/yyyy. Then, it will put that input through a series of if statements and while loops to verify the correct date information. After, it will crunch the number to a single digit, run it through case statements, and print out the corresponding numerology.