Create Program12
This program will take input from the user in the form of an octal number. The max amount of digits that the user is able to enter is 8 digits, and the minimum amount is 0. If the digit is not in that range, the user will be prompted to re-enter the number. Once they user inputs that number, the program will convert the octal number back into the decimal form.